[feat] improve find function
