--- src/main/java/org/example/schedulemanagementapp/ScheduleService.java
@@ -1,44 +1,0 @@
-package org.example.schedulemanagementapp;
-
-import lombok.RequiredArgsConstructor;
-import org.example.schedulemanagementapp.dto.ScheduleRequestDto;
-import org.example.schedulemanagementapp.dto.ScheduleResponseDto;
-import org.example.schedulemanagementapp.entity.Schedule;
-import org.example.schedulemanagementapp.repository.ScheduleRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class ScheduleService {
-
-    private final ScheduleRepository scheduleRepository;
-
-    @Transactional
-    public ScheduleResponseDto save(ScheduleRequestDto scheduleRequestDto) {
-        Schedule schedule = new Schedule(
-                scheduleRequestDto.getTitle(),
-                scheduleRequestDto.getContents(),
-                scheduleRequestDto.getName(),
-                scheduleRequestDto.getPassword()
-        );
-        Schedule savedMember = scheduleRepository.save(schedule);
-
-        return new ScheduleResponseDto(savedMember.getId(),savedMember.getTitle(),savedMember.getContents(), savedMember.getName(), savedMember.getCreatedAt(), savedMember.getModifiedAt());
-    }
-
-    @Transactional(readOnly = true)
-    public List<ScheduleResponseDto> findAll() {
-        List<Schedule> scheduleList = scheduleRepository.findAll();
-
-        List<ScheduleResponseDto> scheduleResponseDtoList = new ArrayList<>();
-
-        for (Schedule schedule : scheduleList) {
-            scheduleResponseDtoList.add(new ScheduleResponseDto(schedule.getId(), schedule.getTitle(), schedule.getContents(), schedule.getName(), schedule.getCreatedAt(), schedule.getModifiedAt()));
-        }
-        return scheduleResponseDtoList;
-    }
-}
