+++ src/main/java/org/example/schedulemanagementapp/controller/ScheduleController.java
@@ -1,13 +1,10 @@
 package org.example.schedulemanagementapp.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.example.schedulemanagementapp.ScheduleService;
+import org.example.schedulemanagementapp.service.ScheduleService;
 import org.example.schedulemanagementapp.dto.ScheduleRequestDto;
 import org.example.schedulemanagementapp.dto.ScheduleResponseDto;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -22,7 +19,7 @@ public class ScheduleController {
     }
 
     @GetMapping("/schedule")
-    public List<ScheduleResponseDto> getMembers() {
-        return scheduleService.findAll();
+    public List<ScheduleResponseDto> getMembers(@RequestParam(required = false) String name) {
+        return scheduleService.findAll(name);
     }
 }
